🔧 More wrenchability
Allowing blocks to be picked up with a wrench via the IWrenchable interface instead of going through a tag feels more consistent with how Create does it with its own blocks. This also allows some blocks to be rotated with the wrench, i.e. the Catalytic Converter and the Pollutometer.