⚡ Fixed a couple wrench related issues with the Dynamo when in arc furnace mode
This fixes the lid of the arc furnace turning into a second dynamo when the actual dynamo is right clicked with a wrench instead of rotating as intended. The dynamo also didn't revert back to its normal form when the lid is removed, so I've added an extra check to address that.