↩️ Undoing Vat mole display fix for now
Undoing a fix for Vats displaying the wrong number of moles when there is no gas phase present because I'm kind of iffy about it.

The fix involved changing VatFluidTankBehaviour.getCombinedReadOnlyMixture to return a mixture with an implied volume of 1 mB, meaning concentration = number of moles, which works, but feels kind of wrong for what the function implies it does.

The problem comes from getCombinedReadOnlyMixture returning a mixture with an implied volume equal to the combined volume of both the liquid and gas phases, but the concentrations from this mixtures are then multiplied by the vat's total capacity when displaying the number of moles in the GUI. This is fine unless there is no gas phase in the Vat, in which case the vat's capacity is no longer equal to the combined volume. A more "correct" fix might be to ignore the volume of the liquid and gas phases entirely and assume the returned mixture will always take up the entire vat, but I'm not entirely sure so I'm leaving it be for now.