🧂 Salt ingredient fix (?)
Saw this in today's fixes and I think the cation and anion charges need to be switched around when calculating the required concentrations? 1M FeCl3 is 1M Fe3+ and 3M Cl- if I'm not mistaken.
I've also switched the gcd function to the one from IntMath, which is more optimized.